# docs/swagger.yaml
basePath: /
definitions:
  api.PUTEdgesReq:
    properties:
      edges:
        items:
          $ref: '#/definitions/models.Edge'
        type: array
    type: object
  api.PUTNodesReq:
    properties:
      nodes:
        items:
          $ref: '#/definitions/models.Node'
        type: array
    type: object
  api.UploadReq:
    properties:
      edges:
        items:
          $ref: '#/definitions/models.Edge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/models.Node'
        type: array
    type: object
  api.UploadedResp:
    properties:
      edges:
        items:
          $ref: '#/definitions/models.Edge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/models.Node'
        type: array
    type: object
  models.Edge:
    properties:
      created_at:
        type: string
      from_id:
        type: integer
      id:
        type: integer
      label:
        type: string
      properties:
        $ref: '#/definitions/models.Properties'
      snippet:
        description: this is a special field show a small snippet of the match terms
        type: string
      to_id:
        type: integer
      updated_at:
        type: string
      weight:
        type: integer
    type: object
  models.Graph:
    properties:
      links:
        items:
          $ref: '#/definitions/models.Edge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/models.Node'
        type: array
    type: object
  models.Node:
    properties:
      created_at:
        type: string
      id:
        type: integer
      label:
        type: string
      properties:
        $ref: '#/definitions/models.Properties'
      snippet:
        description: this is a special field show a small snippet of the match terms
        type: string
      updated_at:
        type: string
    type: object
  models.Properties:
    additionalProperties: {}
    type: object
info:
  contact: {}
  description: EdgeDB API server
  title: EdgeDB API
  version: "1.0"
paths:
  /api/v1/edges:
    get:
      description: Search and return edges.
      parameters:
      - default: ""
        description: search term
        in: query
        name: term
        type: string
      - default: <span class="text-red-500">
        description: snippet start
        in: query
        name: snippetStart
        type: string
      - default: </span>
        description: snippet start
        in: query
        name: snippetEnd
        type: string
      - default: 10
        description: snippet tokens
        in: query
        maximum: 64
        minimum: 1
        name: tokens
        type: integer
      - default: 1000
        description: limit results returned
        in: query
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of edges
          schema:
            items:
              $ref: '#/definitions/models.Edge'
            type: array
        "400":
          description: Bad request
        "500":
          description: Internal server error
      summary: Search and return edges
      tags:
      - edges
    put:
      description: Add/update on or more edges.
      parameters:
      - description: One or more nodes to add/update
        in: body
        name: nodes
        required: true
        schema:
          $ref: '#/definitions/api.PUTEdgesReq'
      produces:
      - application/json
      responses:
        "200":
          description: List of edges
          schema:
            items:
              $ref: '#/definitions/models.Edge'
            type: array
        "400":
          description: Bad request
        "500":
          description: Internal server error
      summary: Add/update one or more edges.
      tags:
      - edges
  /api/v1/graph:
    get:
      description: search return nodes and edges in a format that can be used in a
        force directed graph
      parameters:
      - default: ""
        description: search term
        in: query
        name: term
        type: string
      - default: <span class="text-red-500">
        description: snippet start
        in: query
        name: snippetStart
        type: string
      - default: </span>
        description: snippet start
        in: query
        name: snippetEnd
        type: string
      - default: 10
        description: snippet tokens
        in: query
        maximum: 64
        minimum: 1
        name: tokens
        type: integer
      - default: 1000
        description: limit results returned
        in: query
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Payload used for drawing graphs.
          schema:
            $ref: '#/definitions/models.Graph'
        "400":
          description: Bad request
        "500":
          description: Internal server error
      summary: search return nodes and edges used for a force directed graph
      tags:
      - graph
  /api/v1/nodes:
    get:
      description: Search and return nodes
      parameters:
      - default: ""
        description: search term
        in: query
        name: term
        type: string
      - default: <span class="text-red-500">
        description: snippet start
        in: query
        name: snippetStart
        type: string
      - default: </span>
        description: snippet start
        in: query
        name: snippetEnd
        type: string
      - default: 10
        description: snippet tokens
        in: query
        maximum: 64
        minimum: 1
        name: tokens
        type: integer
      - default: 1000
        description: limit results returned
        in: query
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of nodes
          schema:
            items:
              $ref: '#/definitions/models.Node'
            type: array
        "400":
          description: Bad request
        "500":
          description: Internal server error
      summary: Search and return nodes
      tags:
      - nodes
    put:
      description: Add/update on or more nodes.
      parameters:
      - description: One or more nodes to add/update
        in: body
        name: nodes
        required: true
        schema:
          $ref: '#/definitions/api.PUTNodesReq'
      produces:
      - application/json
      responses:
        "200":
          description: List of nodes
          schema:
            items:
              $ref: '#/definitions/models.Node'
            type: array
        "400":
          description: Bad request
        "500":
          description: Internal server error
      summary: Add/update one or more nodes.
      tags:
      - nodes
  /api/v1/upload:
    put:
      description: Uploads one or more nodes and edges sets.
      parameters:
      - description: One or more nodes to add/update
        in: body
        name: nodes
        required: true
        schema:
          $ref: '#/definitions/api.UploadReq'
      produces:
      - application/json
      responses:
        "200":
          description: List of uploaded nodes and edges
          schema:
            items:
              $ref: '#/definitions/api.UploadedResp'
            type: array
        "400":
          description: Bad request
        "500":
          description: Internal server error
      summary: Uploads one or more nodes and edges sets.
      tags:
      - upload
  /static:
    get:
      description: Static serves up static files
      responses: {}
      summary: Static serves up static files
swagger: "2.0"